AI_Assistant/backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
chromadb
sentence-transformers
openpyxl
python-dotenv
pydantic
langchain-community
pypdf
azure-storage-blob
langchain-huggingface
huggingface_hub[hf_xet]